Sprites

+++ src/Quarter_4/NewSideScroller/BackgroundLayer.java
@@ -5,11 +5,11 @@ import javafx.scene.image.Image;
 public class BackgroundLayer {
 
     private Image image;
-    private int position;
-    private int width;
-    private int height;
+    private double position;
+    private double width;
+    private double height;
 
-    public BackgroundLayer(Image image, int position, int width, int height) {
+    public BackgroundLayer(Image image, double position, double width, double height) {
         this.image = image;
         this.position = position;
         this.width = width;
@@ -19,12 +19,12 @@ public class BackgroundLayer {
     public Image getImage() { return image; }
     public void setImage(Image image) { this.image = image; }
 
-    public int getPosition() { return position; }
-    public void setPosition(int position) { this.position = position; }
+    public double getPosition() { return position; }
+    public void setPosition(double position) { this.position = position; }
 
-    public int getWidth() { return width; }
-    public void setWidth(int width) { this.width = width; }
+    public double getWidth() { return width; }
+    public void setWidth(double width) { this.width = width; }
 
-    public int getHeight() { return height; }
-    public void setHeight(int height) { this.height = height; }
+    public double getHeight() { return height; }
+    public void setHeight(double height) { this.height = height; }
 }
